transfering class SCell from Rule4 to class itself
